- Rework time stamping to try to unify things a bit. - Switch to using ticks instead of milliseconds in net packets and net logic to improve sub-frame accuracy. - New server time offset logic which takes into account the frame-time delay between draining the incoming packets.

+++ PolyPlane/GameObjects/FighterPlane.cs
@@ -508,7 +508,7 @@ namespace PolyPlane.GameObjects
                 _engineFireFlame.StopSpawning();
         }
 
-        public override void NetUpdate(D2DPoint position, D2DPoint velocity, float rotation, double frameTime)
+        public override void NetUpdate(D2DPoint position, D2DPoint velocity, float rotation, long frameTime)
         {
             base.NetUpdate(position, velocity, rotation, frameTime);
